+++ proto/anki/backend.proto
@@ -25,6 +25,7 @@ enum ServiceIndex {
   SERVICE_INDEX_COLLECTION = 13;
   SERVICE_INDEX_CARDS = 14;
   SERVICE_INDEX_LINKS = 15;
+  SERVICE_INDEX_IMPORT_EXPORT = 16;
 }
 
 message BackendInit {
