+++ rslib/src/collection/mod.rs
@@ -2,7 +2,6 @@
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
 pub mod backup;
-pub mod exporting;
 pub(crate) mod timestamps;
 mod transact;
 pub(crate) mod undo;
