+++ rslib/src/import_export/package/mod.rs
@@ -1,0 +1,11 @@
+// Copyright: Ankitects Pty Ltd and contributors
+// License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
+
+mod colpkg;
+mod meta;
+
+pub(crate) use colpkg::export::export_colpkg_from_data;
+pub use colpkg::import::import_colpkg;
+pub(self) use meta::{Meta, Version};
+
+pub(self) use crate::backend_proto::{media_entries::MediaEntry, MediaEntries};
