+++ rslib/src/lib.rs
@@ -18,6 +18,7 @@ pub mod decks;
 pub mod error;
 pub mod findreplace;
 pub mod i18n;
+pub mod import_export;
 pub mod latex;
 pub mod links;
 pub mod log;
